# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sinal AI Sales OS API
  version: 1.0.0
  description: Internal API for the AI-first SDR MVP. All external side effects are policy-protected and idempotent.
servers:
  - url: https://sales-os.example.internal
security:
  - bearerAuth: []
paths:
...
  /api/enrollments:
    post:
      operationId: enrollContact
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EnrollmentRequest'}}}}
      responses: {'201': {description: Enrolled}, '409': {description: Active enrollment conflict}}
...
  /api/approvals/{id}/decision:
    post:
      operationId: decideApproval
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalDecisionRequest'}}}}
      responses: {'200': {description: Decision recorded}, '409': {description: Approval already decided/expired}}
...
  /api/inbound:
    post:
      operationId: normalizedInbound
      security: [{webhookSignature: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/InboundWebhook'}}}}
      responses: {'202': {description: Accepted}, '409': {description: Duplicate event}}
...
  /api/digests/daily:
    post:
      operationId: generateDailyDigest
      requestBody: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/DailyDigestRequest'}}}}
      responses: {'201': {description: Generated, content: {application/json: {schema: {$ref: '#/components/schemas/DailyDigestResponse'}}}}}
    get:
      operationId: getDailyDigest
      parameters:
        - {name: date, in: query, schema: {type: string, format: date}}
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/DailyDigestResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    webhookSignature: {type: apiKey, in: header, name: X-SalesOS-Signature}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    NotFound: {description: Not found}
  schemas:
...
    EnrollmentRequest:
      type: object
      required: [campaignId,contactId]
      properties: {campaignId: {type: string, format: uuid}, contactId: {type: string, format: uuid}}
...
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties: {decision: {enum: [APPROVE,EDIT,REJECT]}, editedPayload: {type: object}, reason: {type: string}}
...
    DailyDigestRequest:
      type: object
      properties: {date: {type: string, format: date}}
    DailyDigestResponse:
      type: object
      required: [digestId,content]
      properties: {digestId: {type: string, format: uuid}, content: {type: object}, generatedAt: {type: string, format: date-time}}
    InboundWebhook:
      type: object
      required: [provider,eventId,channel,address,receivedAt,rawContent]
      properties: {provider: {type: string}, eventId: {type: string}, channel: {type: string}, address: {type: string}, receivedAt: {type: string, format: date-time}, rawContent: {type: string}, providerThreadId: {type: string}}
...